requirements.txt
acres==0.3.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.8
cloudinary==1.38.0
configobj==5.0.9
configparser==7.2.0
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
deepmerge==1.1.1
Deprecated==1.2.14
dj-database-url==2.1.0
Django==4.2.7
django-appconf==1.0.6
django-compressor==4.4
django-environ==0.11.2
django-paypal==2.0
dotmap==1.3.30
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.18.0
firebase-admin==6.4.0
fonttools==4.47.0
frozenlist==1.4.1
gcloud==0.18.3
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-genai==1.8.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.6
importlib_resources==6.5.2
isodate==0.6.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.1
kiwisolver==1.4.5
looseversion==1.3.0
lxml==5.3.1
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
msgpack==1.0.7
multidict==6.1.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==1.26.3
oauth2client==4.1.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pathlib==1.0.1
pillow==10.2.0
propcache==0.2.0
proto-plus==1.23.0
protobuf==4.25.2
prov==2.0.1
psycopg2-binary==2.9.9
puremagic==1.28
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.11.0
pydantic_core==2.33.0
pydot==3.0.4
Pygments==2.17.2
PyJWT==2.8.0
PyMuPDF==1.25.4
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-jwt==4.1.0
pytz==2023.3.post1
pyxnat==1.6.3
pyzbar==0.1.9
rcssmin==1.1.1
rdflib==6.3.2
referencing==0.32.1
requests==2.29.0
requests-toolbelt==0.10.1
rich==13.7.0
rjsmin==1.2.1
rpds-py==0.16.2
rsa==4.9
scipy==1.15.2
screeninfo==0.8.1
simplejson==3.20.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.5
traits==7.0.2
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
websockets==15.0.1
whitenoise==6.6.0
wrapt==1.16.0
yarl==1.15.5
